Rename xml_parser to file_io.
